# Optimize SQL queries for fetching product variants