--- a/saleor/product/managers.py
+++ b/saleor/product/managers.py
@@ -306,7 +306,7 @@ class ProductVariantQueryset(models.QuerySet):
         channel: Optional[Channel],
         limited_channel_access: bool,
     ):
-        from .models import ALL_PRODUCTS_PERMISSIONS, ProductVariant
+        from .models import ALL_PRODUCTS_PERMISSIONS
 
         # User with product permissions can see all variants. If channel is given,
         # filter variants with product channel listings for this channel.
@@ -327,7 +327,7 @@ class ProductVariantQueryset(models.QuerySet):
         # - have a variant channel listing for this channel and the price is not null
         # - have a product channel listing for this channel and the product is published
         #  and visible in listings
-        variants = ProductVariant.objects.using(self.db).filter(
+        variants = self.filter(
             channel_listings__channel_id=channel.id,
             channel_listings__price_amount__isnull=False,
         )